Fix ssl in nginx
The default SSL_ENABLED is off when running the container. However, when choosing to enable it with SSL_ENABLED="yes" after it was ran with no ssl, the ssl didn't work. This fix allows toggling between non-ssl and ssl when stopping/removing and restarting the container.

Also, I deleted the redundant nginx_ssl.conf file, since the only difference with nginx.conf was the listening port; I enabled toggling the listening port with a sed command.